use statsd instead of dogstatsd for sirun
